Update Look of Character Module Update the icon and removed details from main cards to be shown in modal.
