+++ src/Components/Styling/Character.css
@@ -1,0 +1,3 @@
+.smallIcon {
+    height: 20px;
+}
