+++ src/Components/Styling/Modal.css
@@ -1,0 +1,20 @@
+.modal {
+    position: fixed;
+    border: 3px solid white;
+    background-color: black;
+    box-shadow: -2rem 2rem 2rem rgba(white, 0.2);
+    opacity: 0.95;
+    border-radius: 12px;
+    padding: 1rem;
+    z-index: 1;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+
+    
+  }
+  
+  .modal-close {
+    float: right;
+  }
+
